Extract Prompt Studio advanced value sync
